Add QueryBuilder component to QueryEditor Fixes #1155
