+++ ui/src/constants/editorShapes/index.ts
@@ -4,6 +4,6 @@ import ValueSetShape from "./ValueSet";
 import DataModelShape from "./DataModel";
 import FolderShape from "./Folder";
 import PropertyShape from "./Property";
-import QueryShape from "./Query";
+import CohortQueryShape from "./CohortQuery";
 
-export default [ConceptShape, ConceptSetShape, ValueSetShape, FolderShape, DataModelShape, PropertyShape, QueryShape];
+export default [ConceptShape, ConceptSetShape, ValueSetShape, FolderShape, DataModelShape, PropertyShape, CohortQueryShape];
